ostroproject-ci.inc: only build supported recipes
From now on we treat it as an error when a pull request causes a build
to include new components. Those components must be accepted into
Ostro OS first by checking whether we can support them and then
extending the ostro-supported-recipes.txt file.

That is necessary because we must not release binaries that we cannot
support later one.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>