ostro-os: enable and document schroot/debootstrap
This adds a recipe for schroot which works on Ostro OS (not packaged
elsewhere and has to be adapted to the distro) and explains how to use
that in combination with debootstrap (available in
meta-openembedded/meta-oe).

The goal is to make it possible to run Debian or Ubuntu binaries
unmodified on Ostro OS, for those cases where tools are not available
for Ostro OS itself. Compiling inside that chroot environment also
works, but the resulting binaries then will also only work in that
environment.

The content of Ostro images does not get modified. Developers are
currently expected to do that themselves in their custom images; if
deemed useful, the tools can be pre-installed in a separate commit.

Other approaches were also considered:
* creating the rootfs with Docket is more complicated (additional
  dependencies, did not compile immediately for Ostro OS)
* running commands inside the chroot via systemd-nspawn or the AppFW
  would also work, but is harder to configure and more suitable for
  actual services

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>